+++ myokit/tests/shared.py
@@ -31,7 +31,10 @@ DIR_IO = os.path.join(DIR_DATA, 'io')
 DIR_FORMATS = os.path.join(DIR_DATA, 'formats')
 
 # OpenCL support
-OpenCL_FOUND = myokit.OpenCL.supported()
+OpenCL_FOUND = myokit.OpenCL.supported_and_available()
+OpenCL_DOUBLE_PRECISION_CONNECTIONS = (
+    OpenCL_FOUND and myokit.OpenCL.test_extension_on_current_platform(
+        'cl_khr_int64_base_atomics'))
 
 
 class TemporaryDirectory(object):